Major Renaming Of Files
Renamed most of the files. I'm trying to let GitHub take care of version
control and stop using my previous naming convention.
The "Eddie.spin" file is the top object when using the original Eddie
hardware. The "EddieActivityBoard.spin" file is the top object for use
with a Propeller Activity Board and HB-25 motor controllers.