Added WS2812B LEDs to Experimental Version
I'm experimenting with using WS2812B LEDs as a form or visual feedback.
I've added the commands "BRT" which sets the overall brightness of the
LEDs. The command "COLOR" sets the color of a single LED and and the
command "COLORS" sets the color of multiple LEDs. See OPP#8 thread for
more information.